+++ internal/app/root_command.go
@@ -26,7 +26,14 @@ func newRootCommand(resources *applicationResources) *cobra.Command {
 		},
 	}
 
-	configureServeFlags(rootCommand.Flags(), resources.configurationManager, true)
+	serveFlags := pflag.NewFlagSet("serve", pflag.ContinueOnError)
+	configureServeFlags(serveFlags, resources.configurationManager)
+	rootCommand.Flags().AddFlagSet(serveFlags)
+
+	httpsOptionFlags := pflag.NewFlagSet("serve-https-options", pflag.ContinueOnError)
+	configureServeHTTPSOptions(httpsOptionFlags, resources.configurationManager)
+	rootCommand.Flags().AddFlagSet(httpsOptionFlags)
+
 	rootCommand.Flags().String(flagNameTLSCertificatePath, resources.configurationManager.GetString(configKeyServeTLSCertificatePath), "Path to TLS certificate (PEM)")
 	rootCommand.Flags().String(flagNameTLSKeyPath, resources.configurationManager.GetString(configKeyServeTLSKeyPath), "Path to TLS private key (PEM)")
 	_ = resources.configurationManager.BindPFlag(configKeyServeTLSCertificatePath, rootCommand.Flags().Lookup(flagNameTLSCertificatePath))
@@ -34,12 +41,12 @@ func newRootCommand(resources *applicationResources) *cobra.Command {
 
 	rootCommand.PersistentFlags().String(flagNameConfigFile, "", "Path to configuration file")
 
-	rootCommand.AddCommand(newHTTPSCommand(resources))
+	rootCommand.AddCommand(newHTTPSCommand(resources, serveFlags, httpsOptionFlags))
 
 	return rootCommand
 }
 
-func configureServeFlags(flagSet *pflag.FlagSet, configurationManager *viper.Viper, includeHTTPSOptions bool) {
+func configureServeFlags(flagSet *pflag.FlagSet, configurationManager *viper.Viper) {
 	flagSet.String(flagNameBindAddress, configurationManager.GetString(configKeyServeBindAddress), "Specify bind address")
 	flagSet.String(flagNameDirectory, configurationManager.GetString(configKeyServeDirectory), "Serve files from this directory")
 	flagSet.String(flagNameProtocol, configurationManager.GetString(configKeyServeProtocol), "HTTP protocol version (HTTP/1.0 or HTTP/1.1)")
@@ -52,10 +59,11 @@ func configureServeFlags(flagSet *pflag.FlagSet, configurationManager *viper.Vip
 	_ = configurationManager.BindPFlag(configKeyServeNoMarkdown, flagSet.Lookup(flagNameNoMarkdown))
 	_ = configurationManager.BindPFlag(configKeyServeBrowse, flagSet.Lookup(flagNameBrowse))
 	_ = configurationManager.BindPFlag(configKeyServeLoggingType, flagSet.Lookup(flagNameLoggingType))
-	if includeHTTPSOptions {
-		flagSet.Bool(flagNameHTTPS, configurationManager.GetBool(configKeyServeHTTPS), "Serve over HTTPS using a self-signed certificate")
-		flagSet.StringSlice(flagNameHTTPSHosts, configurationManager.GetStringSlice(configKeyHTTPSHosts), "Hostnames or IP addresses for automatic HTTPS certificates")
-		_ = configurationManager.BindPFlag(configKeyServeHTTPS, flagSet.Lookup(flagNameHTTPS))
-		_ = configurationManager.BindPFlag(configKeyHTTPSHosts, flagSet.Lookup(flagNameHTTPSHosts))
-	}
+}
+
+func configureServeHTTPSOptions(flagSet *pflag.FlagSet, configurationManager *viper.Viper) {
+	flagSet.Bool(flagNameHTTPS, configurationManager.GetBool(configKeyServeHTTPS), "Serve over HTTPS using a self-signed certificate")
+	flagSet.StringSlice(flagNameHTTPSHosts, configurationManager.GetStringSlice(configKeyHTTPSHosts), "Hostnames or IP addresses for automatic HTTPS certificates")
+	_ = configurationManager.BindPFlag(configKeyServeHTTPS, flagSet.Lookup(flagNameHTTPS))
+	_ = configurationManager.BindPFlag(configKeyHTTPSHosts, flagSet.Lookup(flagNameHTTPSHosts))
 }